Setting timeout checks in connect/6 using set_stream/2.
This simplifies the whole process greatly by removing the message queues
associated with timers for ping checks, etc. Using this method, the
conenction will stay alive so long as activity is monitored on the
socket stream. We should now be able to use the protocol without making
calls to thread_send_message(tq, true), etc.